requirements.txt
tensorflow==2.16.1
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
pillow==11.1.0
opencv-python-headless==4.10.0.84
mlflow==2.20.3
fastapi==0.115.8
uvicorn[standard]==0.34.0
streamlit==1.43.2
pyyaml==6.0.2
matplotlib==3.10.1
python-multipart==0.0.20
pytest==8.3.5
httpx==0.28.1
